requirements.txt
alembic==1.11.1
numpy==1.21.3
pandas==1.3.3
SQLAlchemy==1.4.32
SQLAlchemy_Utils==0.41.1
pandas==1.3.3
kaggle==1.5.13
psycopg2_binary==2.9.4
openpyxl==3.1.2
python-dotenv==1.0.0